backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
asyncpg
sqlalchemy
scikit-learn
pandas
joblib
numpy
firebase-admin
